functions/mcp_cosmos/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cosmos>=4.7,<5.0
azure-identity>=1.15,<2.0
